requirements.txt
beautifulsoup4==4.13.3
python-dotenv==1.0.1
requests==2.32.3
selenium==4.30.0
seleniumbase==4.36.4
serpapi==0.1.5
